backend/pyproject.toml
[project]
name = "hannaleon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.125.0",
    "bcrypt==4.0.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
